Switch to GitHub Actions to resume builds (#1)
Travis stopped building images on their .org instance on 15 Jun 2021,
leaving the last push to Docker Hub on 19 May 2021. That's a few months
out of date, now, which causes pulls to have lots of updates to run on
startup. In the interest of getting builds going again, and moving away
from Travis, here is a switch to GitHub Actions instead. I also replaced
the QEMU setup script with a Docker image designed to do the same thing,
mostly so the builds would complete.